backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
requests
python-dotenv
python-multipart
bcrypt==4.0.1

